% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{bam_covariate_importance_v5}
\alias{bam_covariate_importance_v5}
\title{BAM Covariate Importance Data (v5)}
\format{
A `data.frame` with 7 columns:
\describe{
  \item{spp}{A four-letter bird code indicating the species.}
  \item{bcr}{The Bird Conservation Region.}
  \item{var}{The name of the covariate.}
  \item{mean_rel_inf}{The mean relative influence of the covariate across bootstraps.}
  \item{sd_rel_inf}{The standard deviation of the covariate's relative influence across bootstraps.}
  \item{n_boots}{The number of bootstraps in which the given covariate appeared.}
  \item{var_class}{A broad variable class to which the covariate belongs.}
}
}
\source{
The data was derived from internal project datasets.
}
\usage{
bam_covariate_importance_v5
}
\description{
`bam_covariate_importance_v5` is a `data.frame` where each row represents the mean relative influence (calculated from different number of  bootstraps) of a model covariate for a given species × Bird Conservation Region (BCR).
}
\examples{
data(bam_covariate_importance_v5)
head(bam_covariate_importance_v5)
}
\keyword{datasets}
