% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_occurrence.R
\name{bam_occurrence}
\alias{bam_occurrence}
\title{Estimate a presence/absence threshold using a Lorenz curve}
\usage{
bam_occurrence(raster_list, quantile = "by_lorenz", plot = TRUE)
}
\arguments{
\item{raster_list}{A raster of the "Area of Interest". Defined and created via Melina's function(s).}

\item{quantile}{Default is \code{"by_lorenz"}, and the  optimum threshold is estimated via
\code{opticut::lorenz()} as the pixel density when the slope of the tangent of the Lorenz function is 1.
If a custom threshold is preferred, set this argument using a \code{numeric} between 0 and 1.
This value indicates a cumulative proportion of pixels, above which all raster pixel values are assigned as "presence".
Raster pixel values below the quantile are assigned "absent". E.g. By setting \code{quantile=0.8},
the threshold density for separating presence versus absence is whatever pixel value accumulates 80% of the total
pixels from the raster.}

\item{plot}{Default is \code{TRUE} for visualizing occurrence patterns.}
}
\value{
A list with:
\describe{
  \item{raster}{A \code{SpatRaster} with binary presence (1) / absence (0) values.}
  \item{occurrence_summary}{A \code{data.frame} with estimated area of occupancy
  in km², before and after thresholding.}
  }
}
\description{
This function fits a Lorenz curve to a raster's density
values using \code{opticut::lorenz()}.
The Lorenz curve is not plotted. See package \code{opticut} for
inspecting Lorenz curves. Instead, the plot generated by \code{plot=TRUE} is the user's
input density raster re-coded to presence (1) / absence (0) values.
}
\details{
\code{opticut::lorenz()} generates a Lorenz plot where the x-axis represents the cumulative
proportion of area (sorted by bird density, \code{p}) and the y-axis represents the cumulative
proportion of the raster's total population (\code{L}). If bird density were uniform across
space, the plot would follow a line with a slope of 1. In reality birds are not distributed evenly
on the landscape. \code{lorenz()} estimates a pixel density threshold separating "presence" from "absence".
The threshold is defined as the point on the Lorenz curve where the tangent has a slope of 1 and \code{p - L}
is maximized. See \code{?opticut::lorenz} for details.
Note that in some cases, the Lorenz-derived threshold is similar to the mean density, however, they
estimate different ecological quantities and are not inherently equal.
}
\examples{

# download v4 rasters for Tennessee Warbler and Ovenbird
rasters <- bam_get_layer("v4", c("TEWA", "OVEN"), destfolder=tempdir())

# visualize core habitat
bam_occurrence(rasters)

# set custom threshold and compare core habitat
bam_occurrence(rasters, quantile=0.8)

# analyse core habitat in a custom area, using v5 predictions for Tennessee Warbler
aoi_sf <- vect(system.file("extdata", "vignette_poly_3978.shp", package = "BAMexploreR"))
rasterv5 <- bam_get_layer("v5",  "TEWA", crop_ext = aoi_sf, destfolder = tempdir(), year = "2020")
bam_occurrence(rasterv5)


}
