% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_spp_list.R
\docType{methods}
\name{bam_spp_list}
\alias{bam_spp_list}
\title{Retrieve list of species}
\usage{
bam_spp_list(version, type = "speciesCode", guild = NULL)
}
\arguments{
\item{version}{Specifies the model release: \code{"v5"} for the current
models or \code{"v4"} for the archived models.}

\item{type}{character; type of output provided in the list, either \code{"speciesCode"}, \code{"commonName"} or \code{"scientificName"}.Default is \code{"speciesCode"}.}

\item{guild}{character; Specifies the guild to filter the species list, based on the classification used in The State of Canada’s Birds Report (Birds Canada, 2024).
By providing a guild (e.g., "Forest Birds"), the function will return only the species available within that specific category.
Accepted guild names include \code{"Forest_Birds"}, \code{"Grassland_Birds"}, \code{"Waterfowl"}, and others as defined in the report.}
}
\value{
Vector of species name
}
\description{
The function produce a character vector of species available to download. To
retrieve available species, the function derive the list using the version, and type of output
}
\examples{
speciesList <- bam_spp_list("v4", "speciesCode")
}
\author{
Melina Houle
}
