% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlayerNM_GD.R
\docType{methods}
\name{getlayerNM_GD}
\alias{getlayerNM_GD}
\title{Download National Model species specific output map raster based on list of species}
\usage{
getlayerNM_GD(spList, version, destfile, ext = NULL, year = NULL)
}
\arguments{
\item{spList}{character. A vector of species to be downloaded.}

\item{version}{character. Represents the version of the national model output to be downloaded.}

\item{destfile}{character. Indicate output path where the downloaded file is saved.}

\item{ext}{SpatVector, SpatExtent, or SpatRaster used to define the extent for the cropping.
Or downloading valid BCR polygons from list, type: mapBCR("v4") or mapBCR("v5")}

\item{year}{character; Specify the year for which the density map were generated. Only in v5.}
}
\value{
Invoked for its side-effect of downloading files to the \code{destfile/} directory.
}
\description{
Download National Model species specific output map raster based on list of species
}
\examples{
bird <- getlayerNM("BAOR", "v4", "mean",  tempfile())

bird <- getlayerNM("BAOR", "v4", destfile = tempdir(), layer = "mean", crop = FALSE)

bird <- getlayerNM("BAOR", "v4", destfile = ".", "mean", crop = FALSE, ext = NULL)


}
\author{
Melina Houle
}
