% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionsNM.R
\name{interactionsNM}
\alias{interactionsNM}
\title{Analyze Two-Way Interactions in a Boosted Regression Model for Species and BCR}
\usage{
interactionsNM(data = boot_pts_reduced_i2, bcr, common_name)
}
\arguments{
\item{data}{An exported nested `list` (see: `data(boot_pts_sorted_i2)`) where the top level elements are species x bcr x 2-way interaction tuples,
and the second-level elements are matrices of the associated average and SD of the bootstrapped prediction spaces}

\item{bcr}{A `character` or `numeric` vector specifying the Bird Conservation Region(s) to query.}

\item{common_name}{A `character` vector specifying the species' common name(s) to query.}
}
\value{
A `data.frame` containing the mean and standard deviation of interaction effects for queried covariate pairs,
sorted by the highest mean interaction effect. The result includes the BCR, species, and covariate pairs.
}
\description{
This function extracts and analyzes the two-way interaction spaces (covariate pairs) from
a set of Boosted Regression Model (GBM) objects for a given species and Bird Conservation Region (BCR).
It calculates the mean interaction effect for each covariate pair and sorts the results by the highest interaction effect.
}
\examples{
# Assuming `boot_pts_reduced_i2` contains the required GBM data:
# Example of querying interactions for Alder Flycatcher in BCRs 11 and 12:
# bamexplorer_interactions(data = boot_pts_reduced_i2, bcr = c("can12", "can11"), common_name = "Alder Flycatcher")

# Plotting the GBM interaction for the lowest and highest mean_y_mean:
# plot.gbm(x=b.i, return.grid = FALSE, i.var = c("SCANFITamarack_5x5", "year"), type="response")

}
