% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapBCR.R
\docType{methods}
\name{mapBCR}
\alias{mapBCR}
\title{Retrieve list of BCR overlaid by the study area}
\usage{
mapBCR(ext)
}
\arguments{
\item{ext}{SpatVector, SpatExtent, or SpatRaster used to define the extent for the cropping.}
}
\value{
Vector of bcr overlaid by study area
}
\description{
Retrieve list of BCR overlaid by the study area
}
\examples{
library(data.table)
dt<- data.table(dataset = c("NFDB"),
               url = c("http://cwfis.cfs.nrcan.gc.ca/downloads/nfdb/fire_poly/current_version/"),
               password= c(NA))
listWebData(dt, datasetName = "NFDB")
}
\author{
Melina Houle
}
