% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapBCR_GD.R
\docType{methods}
\name{mapBCR_GD}
\alias{mapBCR_GD}
\title{Retrieve list of BCR overlaid by the study area}
\usage{
mapBCR_GD(version, ext)
}
\arguments{
\item{version}{character; Indicate the version of the National Model requested. Each version of the
National Model has its url access provided within the package.}

\item{ext}{SpatVector, SpatExtent, or SpatRaster used to define the extent for the cropping.}
}
\value{
List containing vector of bcr that overlay the study area and a map illustrating the overlap.
}
\description{
Retrieve list of BCR overlaid by the study area
}
\examples{
subUnit<- mapBCR("v5")
}
\author{
Melina Houle
}
