% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_sizeNM.R
\name{pop_sizeNM}
\alias{pop_sizeNM}
\title{Estimate population size from a density raster}
\usage{
pop_sizeNM(raster_list)
}
\arguments{
\item{raster_list}{A list of \code{SpatRaster}s. See \code{getlayerNM()} for accessing BAM's raster data.}
}
\value{
A \code{tibble} with five columns: \code{spp}, \code{population_size}, \code{mean_density} (per pixel), \code{sd_density}, \code{n_cells}
}
\description{
This function is essentially a wrapper around \code{terra::values},
but with an adjustment for converting pixel values from males/ha to males/pixel.
It also provides a basic summary of the input raster.
}
\examples{
# download rasters for Tennessee Warbler and Ovenbird
rasters <- getlayerNM(c("TEWA", "OVEN"), "v4", destfile=tempdir())

# get summaries of population size
pop_sizeNM(rasters) # 111 million and 3.5 million, respectively
}
