% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppList_GD.R
\docType{methods}
\name{sppList_GD}
\alias{sppList_GD}
\title{Retrieve list of species}
\usage{
sppList_GD(version, type, guild = NULL)
}
\arguments{
\item{version}{character; Indicate the version of the National Model requested. Each version of the
National Model has its url access provided within the package.}

\item{type}{character; type of output provided in the list, either "species_code", "common_name", "family_name" or "scientific_name".}

\item{guild}{character; Specifies the guild to filter the species list, based on the classification used in The State of Canada’s Birds Report (Birds Canada, 2024).
By providing a guild (e.g., "Forest Birds"), the function will return only the species available within that specific category.
Accepted guild names include "Forest_Birds", "Grassland_Birds", "Waterfowl", and others as defined in the report.}
}
\value{
Vector of species name
}
\description{
The function produce a character vector of species available to download. To
retrieve available species, the function derive the list using the version, and type of output
}
\examples{
speciesList <- sppList("v4", "mean", "species_code", guild)
}
\author{
Melina Houle
}
