% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_get_bcr.R
\docType{methods}
\name{bam_get_bcr}
\alias{bam_get_bcr}
\title{Retrieve the list of BCR overlaid by the study area}
\usage{
bam_get_bcr(version, ext = NULL)
}
\arguments{
\item{version}{A \code{character} specifying which version of the BAM landbird model to use. Valid options are "v4" or "v5".}

\item{ext}{A \code{SpatVector} or \code{SpatRaster} used to define the extent for the cropping. If \code{NULL}, the mosaic
will be used as default.}
}
\value{
Vector of bcr that overlay the study area.
}
\description{
Retrieve the list of BCR overlaid by the study area
}
\examples{
subUnit<- bam_get_bcr("v5")
}
\author{
Melina Houle
}
