% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_get_bcr.R
\docType{methods}
\name{bam_get_bcr}
\alias{bam_get_bcr}
\title{Retrieve list of BCR overlaid by the study area}
\usage{
bam_get_bcr(version, ext = NULL)
}
\arguments{
\item{version}{character; Indicate the version of the National Model requested. Each version of the
National Model has its url access provided within the package.}

\item{ext}{SpatVector, SpatExtent, or SpatRaster used to define the extent for the cropping.}
}
\value{
Vector of bcr that overlay the study area.
}
\description{
Retrieve list of BCR overlaid by the study area
}
\examples{
subUnit<- bam_get_bcr("v5")
}
\author{
Melina Houle
}
