% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_map_bcr.R
\docType{methods}
\name{bam_map_bcr}
\alias{bam_map_bcr}
\title{Map the boundaries of BCR subunits for a specified version.}
\usage{
bam_map_bcr(version, ext = NULL)
}
\arguments{
\item{version}{A \code{character} specifying which version of the National Model to use. Valid options are "v4" or "v5".}

\item{ext}{A \code{SpatVector} or a \code{SpatRaster} used to define an area of interest.}
}
\value{
Map illustrating the BCR and overlap extent if provided.
}
\description{
Map the boundaries of BCR subunits for a specified version.
}
\examples{
subUnit<- bam_map_bcr("v5")
}
\author{
Melina Houle
}
