% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_partial_dependence.R
\name{bam_partial_dependence}
\alias{bam_partial_dependence}
\title{Partial Dependence Plot for a Species x BCR x Predictor}
\usage{
bam_partial_dependence(
  species,
  bcr,
  predictor,
  version = "v5",
  colour = "purple"
)
}
\arguments{
\item{species}{A \code{character} specifying the species to generate a partial dependence plot for.
See \code{data(spp_tbl)} for available species and their spelling.}

\item{predictor}{A \code{character} specifying the model predictor (e.g., temperature, precipitation)
to be visualized in the partial dependence plot. See \code{data(predictor_meta)} for available predictors.}

\item{version}{A \code{character}, either \code{"v4"} or \code{"v5"}.
Defaults to \code{"v5"}.}

\item{colour}{A \code{character} specifying the colour of the fitted spline.
Default is \code{"purple"}.}
}
\value{
A \code{ggplot} showing the partial dependence plot, including the
mean response and an error ribbon representing the 95% confidence interval.
}
\description{
Generates a partial dependence plot displaying
how bird density changes over the range of a model predictor for a given species and BCR.
}
\details{
This function fits a \code{smooth.spline()} to bootstrap replicates of the
bird density versus predictor relationship generated by \code{gbm::plot.gbm()},
with \code{continuous.resolution = 50}.
It then and visualizes the mean response across 32 bootstraps
with 95% confidence interval.
Note that partial dependence plots are generated only for species x BCR x predictor
combinations in the top 75 percentile of relative influence across all models.
Run (e.g. for V5) \code{arrange(bam_predictor_importance_v5, desc(mean_rel_inf))}
to view predictors with meaningful correlation with occurrence for a given species x BCR.
}
\examples{
bam_partial_dependence("ALFL", "can10", "ERAMAP_1km", "v5")


}
