% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_pop_size.R
\name{bam_pop_size}
\alias{bam_pop_size}
\title{Estimate population size from a density raster}
\usage{
bam_pop_size(raster_list, crop_ext = NULL, group = NULL)
}
\arguments{
\item{raster_list}{A list of \code{SpatRaster}s. See \code{bam_get_layer()} for accessing BAM's raster data.}

\item{crop_ext}{SpatVector used to define the extent for the cropping and grouping of population estimates.}

\item{group}{Optional character value of column in SpatVector used for grouping population estimates.}
}
\value{
A \code{tibble} with six columns: \code{group}, \code{total_pop}, \code{mean_density} (per pixel), \code{sd_density}, \code{n_cells}, \code{species}
}
\description{
This function is essentially a wrapper around \code{terra::values},
but with an adjustment for converting pixel values from males/ha to males/pixel.
It also provides a basic summary of the input raster.
}
\examples{
# download rasters for Tennessee Warbler and Ovenbird
rasters <- bam_get_layer(c("TEWA", "OVEN"), "v4", destfile=tempdir())

# get summaries of population size
bam_pop_size(rasters) # 111 million and 3.5 million, respectively
}
