% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_predictor_barchart.R
\name{bam_predictor_barchart}
\alias{bam_predictor_barchart}
\title{Visualize Predictor Importance as Stacked Bar Charts}
\usage{
bam_predictor_barchart(
  species = "all",
  bcr = "all",
  groups = c("spp", "predictor_class"),
  version = "v5",
  plot = TRUE,
  colours = NULL
)
}
\arguments{
\item{species}{A \code{character} specifying the species to filter by. The default is \code{"all"}, which includes all species in the dataset.}

\item{bcr}{A \code{character} specifying the model subregions, or Bird Conservation Regions (BCRs) to filter by. The default is \code{"all"}, which includes all BCRs in the dataset.}

\item{groups}{A \code{character} of two grouping variables for summarising predictor importance.
The first group element is plotted on the x-axis as bins each containing a stacked bar,
while the second group element is shown by fill colours in the stacked bars.
Valid strings are any two of: \code{"spp"} (species), \code{"bcr"} (BCR; model subregion), \code{"predictor"} (model predictor), or \code{"predictor_class"} (model predictor class).
Please see the examples below for a visualization.}

\item{version}{A \code{character}. Defaults to \code{"v5"}. Loads BAM's predictor importance data,
a \code{data.frame} containing predictor importance values, with mean predictor importance as
rows and columns \code{bcr}, \code{species}, \code{predictor_clas}, \code{n_boot}, \code{mean_rel_inf}, and \code{sd_rel_inf}.
\code{"v4"} is also possible but not fully supported for all functions in the first release of this package.}

\item{plot}{A \code{logical} indicating whether to plot the results (\code{TRUE}) or return the processed data (\code{FALSE}).}

\item{colours}{A \code{character} vector of hex codes for the colours to use in the ggplot (optional).
If \code{NULL}, default colours are used.}
}
\value{
A stacked bar chart with the first group element plotted on the x-axis as bins each containing a stacked bar, and the second group element is shown by fill colours in the stacked bars.  If plot = FALSE the processed data is returned as a data.frame.
}
\description{
Creates stacked bar plots showing the proportion of model predictors importance
in predicting bird abundance.
}
\details{
Stacked bars can be grouped by species, predictor class, or
Bird Conservation Region (BCR). For example, grouping by species and predictor class creates
a plot where a stacked bar is created for each species, and each bar is split into the proportion
of predictor importance that each of nine predictor classes contributed, pooled across the specified BCRs.
}
\examples{
# Compare predictor importance (binned by predictor class) for all species in all BCRs
bam_predictor_barchart(species = "all", bcr = "all",  groups = c("spp", "predictor_class"))

# Compare predictor importance (binned by predictor class) in the Prairies (BCRs 11, 6-1, 6-0)
# to the Pacific Coast across (BCR 5) all species
prairies_to_coast <- c("can11", "can60", "can61", "can5")
bam_predictor_barchart(species = "all", bcr = prairies_to_coast, groups=c("bcr", "predictor_class"))

# Compare predictor importance (binned by predictor class) for four
# warbler species in BCR14
warblers <- c("CAWA", "BAWW", "BTNW", "BLBW")
bam_predictor_barchart(species = warblers, bcr = "can14", groups = c("spp", "predictor_class"))

# Compare predictor importance for a single warbler species
# relative to the total influence that predictor had across all warblers.
bam_predictor_barchart(species = warblers, bcr = "can14", groups = c("predictor", "spp"))
}
