% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_predictor_importance.R
\name{bam_predictor_importance}
\alias{bam_predictor_importance}
\title{Plot Side-by-Side Variation in Predictor Importance by Predictor Class}
\usage{
bam_predictor_importance(
  species = "all",
  bcr = "all",
  group = "spp",
  version = "v5",
  plot = TRUE,
  colours = NULL,
  viridis_option = "viridis"
)
}
\arguments{
\item{species}{A \code{character} specifying the species to filter by. The default is \code{"all"}, which includes all species in the dataset.}

\item{bcr}{A \code{character} specifying the Bird Conservation Regions (BCRs) to filter by. The default is \code{"all"}, which includes all BCRs in the dataset.
See \code{bam_map_bcr()} for available BCRs and their spellings.}

\item{group}{A \code{character} specifying the grouping variable for summarizing predictor importance.
Valid strings are \code{"spp"} (species), or \code{"bcr"} (BCR).}

\item{version}{Specifies the model release: \code{"v5"} for the current
models or \code{"v4"} for the archived models. Defaults to \code{"v5"}.
Loads the corresponding predictor-importance \code{data.frame}, with columns
\code{bcr}, \code{species}, \code{predictor_class}, \code{n_boot},
\code{mean_rel_inf}, and \code{sd_rel_inf}.}

\item{plot}{A \code{logical} indicating whether to plot the results (\code{TRUE}) or return the processed data (\code{FALSE}).}

\item{colours}{A \code{character} vector of hex codes or colour names (optional).
If \code{NULL}, default colour-blind friendly palette from \code{viridis} is used.}

\item{viridis_option}{A \code{character}. Default is \code{"viridis"}. See \code{?ggplot2::scale_colour_viridis_d} for more options.}
}
\value{
A ggplot displaying percent predictor importance by predictor class, grouped by the \code{group} argument.
Percent importance is used to allow comparisions across groups that have
differing total predictor importance.
If \code{plot = FALSE} the processed data is returned as a \code{data.frame}.
}
\description{
Plot mean relative predictor importance (with bootstrap variation)
by predictor class, with options to group by BCR (Bird Conservation Region) or species.
}
\details{
Bootstrap variation (per species x BCR) is propagated by
taking the root sum square of standard deviation values.
}
\examples{


# Example of plotting predictor importance for Townsend's Solitaire across all BCRs.
# This is a species with relatively high bootstrap variance.
bam_predictor_importance(species = "TOSO")

# Example of plotting predictor importance for two warbler species from three BCRs,
# using custom colours:
bam_predictor_importance(species = c("BAWW", "CAWA"), group = "spp",
bcr = c("can12", "can13", "can14"),  colours = c("#1f78b4", "#33a02c"))



}
