% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotNM.R
\name{boxplotNM}
\alias{boxplotNM}
\title{Side-by-Side Boxplots Displaying Bootstrap Variation by Variable Class}
\usage{
boxplotNM(
  data = NULL,
  version = "v5",
  species = "all",
  bcr = "all",
  group = NULL,
  plot = FALSE,
  colours = NULL
)
}
\arguments{
\item{data}{Defaults to BAM's covariate importance data but can be replaced with any `data.frame` containing covariate importance values,
with covariates as rows and a column each for `bcr`, `species`, `var_class`, `boot`, and `rel.inf`. If user data has no bootstrap replicates, 
create a `boot` column and fill in the same number (e.g. `1`) for all rows. Leave this argument as `NULL` unless using custom data.}

\item{version}{A `character`. Defaults to `"v5"`. `"v4"` is also possible but not fully supported for all functions in the first release of this package.
If using custom data in the `data` argument, this argument will be ignored.}

\item{species}{A `character` specifying the species to filter by. The default is `"all"`, which includes all species in the dataset.}

\item{bcr}{A `character` specifying the Bird Conservation Regions (BCRs) to filter by. The default is `"all"`, which includes all BCRs in the dataset.}

\item{group}{A `character` specifying the grouping variable(s) for summarizing the covariate importance (e.g., species, BCR, etc.).}

\item{plot}{A `logical` indicating whether to plot the results (`TRUE`) or return the processed data (`FALSE`).}

\item{colours}{A `character` vector of hex codes for the colours to use in the boxplot (optional).
If `NULL`, default colours are used.}
}
\value{
A ggplot object displaying boxplots of relative covariate importance by variable class, grouped by the `group` argument.
If `plot = FALSE`, the processed data is returned instead of the plot.
}
\description{
This function generates side-by-side boxplots that show bootstrap variation of covariate importance by variable class,
with options to group by BCR (Bird Conservation Region), species, and other factors.
}
\examples{


# Example of plotting covariate importance for Alder Flycatcher across all BCRs.
# boxplotNM(group = "species", species = "ALFL", bcr = "all")

# Example of plotting covariate importance for two species in two BCRs, using custom colours:
# boxplotNM(group = "species", species = c("ALFL", "CAWA"), 
bcr = c("BCR12", "BCR13"),  colours = c("#1f78b4", "#33a02c"))



}
