% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlayerNM.R
\docType{methods}
\name{getlayerNM}
\alias{getlayerNM}
\title{Download National Model species specific output map raster based on list of species}
\usage{
getlayerNM(spList, version, destfile, layer = "mean", crop = FALSE, ext = NULL)
}
\arguments{
\item{spList}{A character string. A vector of species to be downloaded.}

\item{version}{A character string. Represents the version of the national model output to be downloaded.}

\item{destfile}{A character string giving the path where the downloaded file is saved.}

\item{layer}{Character; Name of the output layers of interest, either "mean", "sd" or "overextrapolated".}

\item{crop}{Logical. If \code{TRUE}, crop raster layers using \code{ext/}.}

\item{ext}{SpatVector, SpatExtent, or SpatRaster used to define the extent for the cropping.}
}
\value{
Invoked for its side-effect of downloading files to the \code{destfile/} directory.
}
\description{
Download National Model species specific output map raster based on list of species
}
\examples{
bird <- getlayerNM("CAWA", "v4", "mean",  tempfile())


}
\author{
Melina Houle
}
