% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapBCR.R
\docType{methods}
\name{mapBCR}
\alias{mapBCR}
\title{Retrieve list of BCR overlaid by the study area}
\usage{
mapBCR(version, ext = NULL)
}
\arguments{
\item{version}{character; Indicate the version of the National Model requested. Each version of the
National Model has its url access provided within the package.}

\item{ext}{SpatVector, SpatExtent, or SpatRaster used to define the extent for the cropping.}
}
\value{
Map illustrating the BCR and overlap extent if provided.
}
\description{
Retrieve list of BCR overlaid by the study area
}
\examples{
subUnit<- mapBCR("v5")
}
\author{
Melina Houle
}
