% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrenceNM.R
\name{occurrenceNM}
\alias{occurrenceNM}
\title{Estimate a presence/absence threshold using a Lorenz curve}
\usage{
occurrenceNM(raster, quantile = NULL, plot = TRUE)
}
\arguments{
\item{raster}{A raster of the "Area of Interest". Defined and created via Melina's function(s).}

\item{quantile}{Default is \code{NULL}, and the  optimum threshold is estimated via
\code{opticut::lorenz()} as the pixel density when the slope of the tangent of the Lorenz function is 1.
If a custom threshold is preferred, set this argument using a \code{numeric} between 0 and 1.
Indicates a cumulative proportion of pixels, above which all raster pixel values are assigned as "presence".
Raster pixel values below the quantile are assigned "absent". E.g. By setting \code{quantile=0.8},
the threshold density for separating presence versus absence is whatever pixel value accumulates 80% of the total
pixels from the raster.}

\item{plot}{Default is \code{TRUE} for visualizing occurrence patterns.}
}
\value{
A list with:
\describe{
  \item{raster}{A \code{SpatRaster} with binary presence (1) / absence (0) values.}
  \item{threshold}{A \code{numeric} value: the estimated density threshold.}
  \item{population_summary}{A \code{data.frame} with estimated total population size,
  mean, and standard deviation, before and after thresholding.
  See: \code{BAMexploreR::pop_sizeNM()}}
}
}
\description{
This function fits a Lorenz curve to a raster's density values using \code{opticut::lorenz()}.
\code{lorenz()} generates a Lorenz plot where the x-axis represents the cumulative
proportion of area (sorted by bird density, \code{p}) and the y-axis represents the cumulative
proportion of the raster's total population (\code{L}). If bird density were uniform across
space, the plot would follow a line with a slope of 1. In reality birds are not distributed evenly
on the landscape. \code{lorenz()} estimates a
pixel density threshold separating "presence" from "absence". The threshold is defined as
the point on the Lorenz curve where the tangent has a slope of 1 and \code{p - L}
is maximized. See \code{?opticut::lorenz} for details.
}
\details{
This function estimates threshold density from a Lorenz curve,
but the Lorenz plot is not provided. See package \code{opticut} for
inspecting Lorenz curves. Instead, the plot generated by \code{plot=TRUE} is the user's
input density raster re-coded to presence (1) / absence (0) values.
}
\examples{
# find Tennessee Warbler core habitat
tewa <- getlayerNM("TEWA", "v4", destfile = tempdir())
occurrenceNM(tewa$TEWA)
}
