% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependenceNM.R
\name{partial_dependenceNM}
\alias{partial_dependenceNM}
\title{Partial Dependence Plot for Species in a BCR Based on Covariate}
\usage{
partial_dependenceNM(data = boot_pts_sorted, bcr, common_name, covariate)
}
\arguments{
\item{data}{A named list of `data.frames`, where each element contains bootstrap replicates for different combinations of BCRs, species, and covariates.
The key for each element in the list is a combination of `bcr`, `common_name`, and `covariate` (e.g., `"12_BAOR_temp"`).}

\item{bcr}{A `character` or `numeric` specifying the Bird Conservation Region (BCR) for which the partial dependence plot is to be generated.}

\item{common_name}{A `character` specifying the common name of the species for which the partial dependence plot is to be generated.}

\item{covariate}{A `character` specifying the covariate (e.g., temperature, precipitation) whose effect is to be visualized in the partial dependence plot.}
}
\value{
A `ggplot` object showing the partial dependence plot, including the mean response and an error ribbon representing the 95% confidence interval.
If the specified combination of BCR, species, and covariate does not exist in the data, the function returns an error.
}
\description{
This function generates a partial dependence plot (PDP) for a given species, BCR (Bird Conservation Region), and covariate,
showing how the response variable changes over the range of the covariate. The function fits a smooth spline to bootstrap replicates,
aggregates predictions, and visualizes the mean response with error bounds.
}
\examples{
partial_dependenceNM(boot_pts_sorted, bcr = "can10", common_name = "Alder Flycatcher", covariate = "year")

# Assuming `boot_pts_sorted` contains the required bootstrap data:
# Example of creating a partial dependence plot for "BAOR" in BCR 12, covariate "temperature":
# partial_dependenceNM(data = boot_pts_sorted,
                       bcr = "12", common_name = "BAOR",
                       covariate = "temperature")
}
