% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_sizeNM.R
\name{pop_sizeNM}
\alias{pop_sizeNM}
\title{Estimate population size from a density raster}
\usage{
pop_sizeNM(raster)
}
\arguments{
\item{raster}{A \code{SpatRaster}. See \code{getlayerNM()} for accessing BAM's raster data.}
}
\value{
A \code{tibble} with four columns: \code{population_size}, \code{mean_density} (per pixel), \code{sd_density}, \code{n_cells}
}
\description{
This function is essentially a wrapper around \code{terra::values},
but with an adjustment for converting pixel values from males/ha to males/pixel.
It also provides a basic summary of the input raster.
}
\examples{
googledrive::drive_auth(scopes = "https://www.googleapis.com/auth/drive.readonly")
root <- "G:/Shared drives/BAM_NationalModels5"
raster <- terra::rast(x="G:/Shared drives/BAM_NationalModels4/NationalModels4.0/Feb2020/artifacts/CAWA/pred-CAWA-CAN-boot-7.tif")
}
