% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackedbarNM.R
\name{stackedbarNM}
\alias{stackedbarNM}
\title{National Models 5.0 - covariate interpretation}
\usage{
stackedbarNM(
  data = bam_covariate_importance,
  groups = NULL,
  species = "all",
  bcr = "all",
  traits = NULL,
  plot = FALSE,
  colours = NULL
)
}
\arguments{
\item{data}{A convariate data frame with rows as types of covariates and columns denoting relative influence.}

\item{groups}{Grouping variables for summarizing model covariates.}

\item{species}{The species to be summarized. (only the input of type 'common_name' or 'sci_name' accepted)}

\item{bcr}{Bird Conservation Regions to be summarized.}

\item{traits}{Traits for the species, e.g., `avonet` or `acad`.}

\item{plot}{Logical; if `TRUE`, creates a stacked bar plot.}

\item{colours}{Character vector of hex codes for colours.}

\item{export}{Logical; if `TRUE`, exports the underlying data.}
}
\value{
A stacked bar chart or a dataframe if `export = TRUE`.
}
\description{
National Models 5.0 - covariate interpretation
}
\examples{
stackedbarNM(bam_covariate_importance, groups = c("common_name", "var_class"), species = 'all', bcr = 'all', plot = TRUE)
}
