% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackedbarNM.R
\name{stackedbarNM}
\alias{stackedbarNM}
\title{This function plots stacked bar charts.}
\usage{
stackedbarNM(
  species = "all",
  bcr = "all",
  groups = NULL,
  version = "v5",
  plot = TRUE,
  colours = NULL
)
}
\arguments{
\item{species}{A \code{character} specifying the species to filter by. The default is \code{"all"}, which includes all species in the dataset.}

\item{bcr}{A \code{character} specifying the Bird Conservation Regions (BCRs) to filter by. The default is \code{"all"}, which includes all BCRs in the dataset.}

\item{groups}{A \code{character} of two grouping variables for summarising covariate importance.
The first group element is plotted on the x-axis as bins each containing a stacked bar,
while the second group element is shown by fill colours in the stacked bars.
Please see the examples below for a visualization.}

\item{version}{A \code{character}. Defaults to \code{"v5"}. Loads BAM's covariate importance data,
a \code{data.frame} containing covariate importance values, with mean covariate importance as
rows and columns \code{bcr}, \code{species}, \code{var_clas}, \code{n_boot}, \code{mean_rel_inf}, and \code{sd_rel_inf}.
\code{"v4"} is also possible but not fully supported for all functions in the first release of this package.}

\item{plot}{A \code{logical} indicating whether to plot the results (\code{TRUE}) or return the processed data (\code{FALSE}).}

\item{colours}{A \code{character} vector of hex codes for the colours to use in the ggplot (optional).
If \code{NULL}, default colours are used.}
}
\value{
A stacked bar chart.
}
\description{
This function plots stacked bar charts.
}
\examples{
# Compare covariate importance (binned by variable class) for four
# warbler species in BCR14
warblers <- c("CAWA", "BAWW", "BTNW", "BLBW")
stackedbarNM(species = warblers, bcr = "can14", groups = c("spp", "var_class"))

# Compare covariate importance for a single warbler species
# relative to the total influence that covariate had across all warblers.
stackedbarNM(species = warblers, bcr = "can14", groups = c("var", "spp"))
}
